The controller for working with the client and working with the balance are placed in separate classes
